Add files via upload
This is a game where the user is instructed to move their character (using the mouse) around to collect snacks from the refrigerator. As the user collects 7 snacks they will be instructed that the character is ready for a nap in which then the scenery changes to a bedroom and the game ends.